requirements.txt
openvino-genai[all]
optimum[openvino]
nncf
SpeechRecognition
PyAudio
gradio
